requirements.txt
pandas
python-dotenv
requests
licensecheck